feat(mastra-multi-provider): refactor provider options validation and remove unused files
