+++ apps/app/src/features/mastra/utils/provider-options-validation.ts
@@ -18,34 +18,89 @@ export const isProviderNamespacedObject = (value: unknown): boolean => {
 };
 
 /**
- * Shared FE/BE formal validation for the `ai:providerOptions` raw JSON string
- * (R6.2). This single predicate is the source of truth for both the client form
- * (react-hook-form `validate` rule) and the server route (express-validator
- * `.custom`), so client and server accept/reject exactly the same input — and it
- * matches what the runtime resolver actually applies, so a value that passes here
- * is never accepted on save but then silently ignored at chat time.
+ * The classification of a providerOptions textarea value, used to drive the inline
+ * status indicator (valid / invalid + why + where).
  *
- * Contract:
- *   - empty / whitespace-only value => valid; it represents "no provider
- *     options", which the container normalizes to `undefined` on save.
- *   - a provider-namespaced JSON object (see `isProviderNamespacedObject`) => valid;
- *     the semantic validity of the per-provider options is each provider
- *     integration's responsibility, not this form. `{}` is valid.
- *   - anything else => invalid: malformed JSON, OR valid JSON of the wrong shape
- *     (a bare primitive such as `42` / `"x"` / `true` / `null`, an array, or an
- *     object whose value is not itself an option object). The runtime ignores
- *     these, so rejecting them up front gives the admin immediate feedback instead
- *     of a silent no-op on save.
+ * `empty` is reported separately from `valid` so the UI can stay neutral for an
+ * unset value (no "valid JSON" affirmation for an empty field). `syntax-error`
+ * carries a 1-based line/column so the admin can locate the typo; `shape-error`
+ * is well-formed JSON of the wrong shape (not a provider-namespaced object), for
+ * which there is no single offending position.
  */
-export const isValidProviderOptionsJson = (value: string): boolean => {
+export type ProviderOptionsJsonStatus =
+  | { kind: 'empty' }
+  | { kind: 'valid' }
+  | { kind: 'syntax-error'; line: number; column: number }
+  | { kind: 'shape-error' };
+
+/**
+ * Locate a `JSON.parse` syntax error in the source text. V8's `SyntaxError`
+ * message carries a `... at position N` byte offset (newer engines also append
+ * `(line L column C)`); deriving line/column from N against the source keeps the
+ * result stable across Node versions. Falls back to the start of the input when
+ * no position can be parsed.
+ *
+ * In-process V8 only (never sent to MongoDB), so a plain `RegExp` is fine here
+ * — see `.claude/rules/mongodb-regex.md`.
+ */
+const locateSyntaxError = (
+  text: string,
+  error: unknown,
+): { line: number; column: number } => {
+  const message = error instanceof Error ? error.message : String(error);
+  const match = /position (\d+)/.exec(message);
+  if (match == null) {
+    return { line: 1, column: 1 };
+  }
+  const position = Number(match[1]);
+  const before = text.slice(0, position);
+  const lastNewline = before.lastIndexOf('\n');
+  return {
+    line: before.split('\n').length,
+    // column is 1-based: chars after the last newline, or from the start.
+    column: position - lastNewline,
+  };
+};
+
+/**
+ * Classify a providerOptions textarea value for inline feedback: the single
+ * parse-and-check pipeline. The FE/BE validator `isValidProviderOptionsJson` is
+ * derived from this (valid iff `empty` or `valid`), so the on-screen indicator and
+ * the rule that actually blocks save can never disagree on the valid/invalid split;
+ * this function additionally reports the *why* (syntax vs. shape) and the
+ * syntax-error location, which the validator does not need.
+ */
+export const getProviderOptionsJsonStatus = (
+  value: string,
+): ProviderOptionsJsonStatus => {
   if (value.trim() === '') {
-    return true;
+    return { kind: 'empty' };
   }
   let parsed: unknown;
   try {
     parsed = JSON.parse(value);
-  } catch {
-    return false;
+  } catch (error) {
+    return { kind: 'syntax-error', ...locateSyntaxError(value, error) };
   }
-  return isProviderNamespacedObject(parsed);
+  return isProviderNamespacedObject(parsed)
+    ? { kind: 'valid' }
+    : { kind: 'shape-error' };
+};
+
+/**
+ * Shared FE/BE formal validation for the `ai:providerOptions` raw JSON string
+ * (R6.2). The source of truth for both the client form (react-hook-form `validate`
+ * rule) and the server route (express-validator `.custom`), so client and server
+ * accept/reject exactly the same input — and it matches what the runtime resolver
+ * applies, so a value that passes here is never accepted on save but then silently
+ * ignored at chat time.
+ *
+ * Derived from `getProviderOptionsJsonStatus` so the valid/invalid split has ONE
+ * definition: valid iff the value is empty or a provider-namespaced object; a
+ * syntax error, a wrong-shape value (bare primitive / array / non-namespaced
+ * object), etc. are all invalid. `{}` is valid (vacuously).
+ */
+export const isValidProviderOptionsJson = (value: string): boolean => {
+  const { kind } = getProviderOptionsJsonStatus(value);
+  return kind === 'empty' || kind === 'valid';
 };
